requirements.txt
# Core framework
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langgraph-checkpoint>=2.0.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0

# Configuration
pyyaml>=6.0.2

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0

# Dev
python-dotenv>=1.0.1
